+++ proyecto-fin-grado/src/main/java/com/example/proyectofingrado/entity/Medico.java
@@ -16,6 +16,6 @@ import java.util.Date;
 public class Medico {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long codigo;
+    private int codigo;
 
 }
